+++ RELEASES.md
@@ -1,3 +1,8 @@
+# Version 0.3.0 (2020-08-16)
+
+  * Use `thiserror` for `Mif` making `anyhow` optional with `cli` feature.
+  * Fix check of whether word value exceeds width for byte-aligned widths.
+
 # Version 0.2.1 (2020-08-12)
 
   * Allow word width smaller than 8 bits.